PyInstaller化できたのでこれを機にクライアント名からGPU削除（CPU版と同一になってたので）
